Fixing model unit testing
In the upgrade of Pylons and partial shift to Pyramid the testing
framework got broken.

I tried hard to fix the old framework but the complexity of it defeated
me. It also seemed excessive to the actual project's needs.

So this is a rework of all the old model tests onto a new framework.

The new tests are part of the zk project as this is where the models
reside.

Tests are executed by pytest, this seems to be the current tool of
choice. From zookeepr root, `py.test`

There is some pytest magic in zk/tests/model/conftest.py
This automatic include is noted in each test file

factory_boy is used to generate dummy data fixtures in fixtures.py
The library is a bit complex but generally a new fixture can be created
by copying an existing one.
Using this library greatly reduces the complexity of each test.

Tests are performed against a real Postgresql database, as this best
matches the deployed system. However switching the SQLite is a one line
change. Each test occurs within a transaction which is rolled back on
completion. This means no data is actually ever commited and tests
should be able to run in parallel.

The old model table tests were all deleted.
These defined the table in the test file, then used common code to
check that the SQLAlchemy definition matched the test definition.
In essence it was testing SQLAlchemy, so wasn't needed.

The old domain tests included an automatic and custom test portion.
The automatic test inserted some sample data into the table and
deleted in, verifying each step. Again this is just testing SQLAlchemy
so I got rid of the automatic portion.
The domain test files which had no custom tests were deleted.