Convert lca_info.sql into an alembic script
This attempts to migrate data from lca_info.py if it exists.
On a downgrade the database contents are exported to rebuild lca_info.py
Both directions should be reviewed by human, just in case.
Pretty messages are thrown at the user in both instances.

If it doesn't exist we use a default data set (built in).
We also reset the default password because we change the seeds.