URL permission test/audit
This extensive (but not comprehensive) test hits a large number of urls
with each of the different roles that we use.

The response is then compared to a known "good" table to ensure that the
permissions are correct.

In practice the "good" table is our current permission set. This test is
designed to provide a backstop and prevent any inadvertant loosening or
regression in permissions.

It also provides a useful overview of the state of the system.